convert to gradle build project
