+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Fri Mar 20 12:49:29 EDT 2020
+gradle.version=5.2.1
