+++ README.MD
@@ -1,14 +1,26 @@
 # CSCI 2020U Final Project
 
+## Contribution Notes  
+
 > Please create a new branch for your own commits
 
-## Members  
+### Contributors  
 
 - Aurie  
 - Paras Chandhok  
 - Xavier Chanthavong  
 - Matthew Mui  
 - Owen Reid  
+
+### Run Program  
+
+> run server.server  
+> run client.clientgui  
+
+### Dependencies  
+
+> javafx-sdk-11.0.2  
+> java jdk 11+  
 
 ## Program Documentation
