--- chatter/README.md
@@ -1,7 +1,0 @@
-run server, run clientgui
-
-dependencies
-
-javafx-sdk-11.0.2
-
-java jdk 11+
